+++ database/migrations/2025_07_22_173452_create_bank_statement_lines_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
             $table->date('date');
             $table->string('description');
             $table->string('partner_name')->nullable();
-            $table->decimal('amount', 15, 2);
+            $table->unsignedBigInteger('amount');
             $table->boolean('is_reconciled')->default(false);
             $table->foreignId('payment_id')->nullable()->constrained()->nullOnDelete();
             $table->timestamps();
